+++ CHANGELOG.md
@@ -5,8 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Unreleased]: https://github.com/ntrrg/ntgo/compare/v0.3.1...master
+[Unreleased]: https://github.com/ntrrg/ntgo/compare/v0.4.0...master
 ## [Unreleased][]
+
+[0.4.0]: https://github.com/ntrrg/ntgo/compare/v0.3.1...v0.4.0
+## [0.4.0][]
 
 ### Add